reaggle/backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#

dj-database-url==0.4.1
django-cors-headers==1.1.0
django==1.9.7
djangorestframework==3.3.3
gunicorn==19.6.0
psycopg2==2.6.1
whitenoise==3.2
